Rename ContributorRepositoryInterface to ContributorRepository and implementing ContributorRepository to DbalContributorRepository
