+++ app/Offer/OfferServiceProvider.php
@@ -9,7 +9,7 @@ use CultuurNet\UDB3\ApiGuard\Consumer\ConsumerReadRepository;
 use CultuurNet\UDB3\ApiGuard\Consumer\Specification\ConsumerIsInPermissionGroup;
 use CultuurNet\UDB3\Broadway\EventHandling\ReplayFilteringEventListener;
 use CultuurNet\UDB3\Container\AbstractServiceProvider;
-use CultuurNet\UDB3\Contributor\ContributorRepository;
+use CultuurNet\UDB3\Contributor\DbalContributorRepository;
 use CultuurNet\UDB3\Event\ReadModel\Relations\EventRelationsRepository;
 use CultuurNet\UDB3\Http\Offer\AddImageRequestHandler;
 use CultuurNet\UDB3\Http\Offer\AddLabelFromJsonBodyRequestHandler;
@@ -96,7 +96,7 @@ final class OfferServiceProvider extends AbstractServiceProvider
             OfferMetadataProjector::class,
             AutoApproveForUiTIDv1ApiKeysProcessManager::class,
             PopularityRepository::class,
-            ContributorRepository::class,
+            DbalContributorRepository::class,
             'iri_offer_identifier_factory',
             'should_auto_approve_new_offer',
             OfferRepository::class,
@@ -570,7 +570,7 @@ final class OfferServiceProvider extends AbstractServiceProvider
             UpdateContributorsRequestHandler::class,
             fn () => new UpdateContributorsRequestHandler(
                 $container->get(OfferRepository::class),
-                $container->get(ContributorRepository::class)
+                $container->get(DbalContributorRepository::class)
             )
         );
