+++ tests/Http/Organizer/UpdateContributorsRequestHandlerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Http\Organizer;
 
 use Broadway\Repository\AggregateNotFoundException;
-use CultuurNet\UDB3\Contributor\ContributorRepositoryInterface;
+use CultuurNet\UDB3\Contributor\ContributorRepository;
 use CultuurNet\UDB3\Http\ApiProblem\ApiProblem;
 use CultuurNet\UDB3\Http\ApiProblem\AssertApiProblemTrait;
 use CultuurNet\UDB3\Http\ApiProblem\SchemaError;
@@ -39,7 +39,7 @@ final class UpdateContributorsRequestHandlerTest extends TestCase
         $this->organizerRepository = $this->createMock(OrganizerRepository::class);
         $this->updateContributorsRequestHandler = new UpdateContributorsRequestHandler(
             $this->organizerRepository,
-            $this->createMock(ContributorRepositoryInterface::class)
+            $this->createMock(ContributorRepository::class)
         );
 
         $this->psr7RequestBuilder = new Psr7RequestBuilder();
